requirements.txt
# Core Python packages
fastapi==0.116.1
uvicorn[standard]==0.35.0
streamlit==1.49.0
pydantic==2.11.7

# ML/AI packages
sentence-transformers==5.1.0
langchain==0.3.27
langchain-openai==0.3.32
openai>=1.107.0
chromadb==1.0.20
numpy==2.3.2

# Data processing
requests==2.32.5
beautifulsoup4==4.13.5
crawl4ai==0.7.4

# Evaluation and NLP
nltk==3.9.1
evaluate==0.4.5
rouge-score==0.1.2
absl-py==2.3.0
tiktoken==0.11.0

# Utilities
python-dotenv==1.1.1
python-multipart==0.0.20
aiofiles==24.1.0
